pyproject.toml
[project]
name = "deep-research-service"
version = "0.1.0"
description = "Multi-agent deep research service powered by AgentScope with async DAG execution"
requires-python = ">=3.10"
dependencies = [
    "agentscope>=1.0.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
pdf = ["weasyprint>=62.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
